move app.main as otaclient.main, cleanup
